added several representation models in order to download infromation from the web server.  finished network request for searching for group names on behalf of user to return matched groups, but the urlrequest is missing the auth0 authentication piece.  This isn't normally required for iOS bc it's a walled garden whereas for Web developers a hacker could theoretically be copming from ANY url on their service.  So first I have to add Web Dev's version of Auth0 which is cumbersome and time-consuming, and then I can see if i get back any data.  right now the data coming back is an error thanks to no authentication piece in the request.  this is going to take some time.  the networking code tests SAT except for no data is fetched thus I can't check the final piece-- to see if the data successfully goes into my array of Group.  Also added some more code to Profile viewcontroller to at least printto the console if a user presses a button.  obviously added the network call from GroupsExistviewcontroller and this entailed also some code to distinguish if any information was recieved, else the app taps saved CoreData groups to display.Will need to figure out where to put idToken credentials from Auth0 in all network requests from here forward

+++ AllegiancePractise/AllegiancePractise/MODELS/Authentication Views/GroupSearchViewController.swift
@@ -44,6 +44,8 @@ class GroupSearchViewController: UIViewController, UITextFieldDelegate {
         //using favoriteTeamTextField.text locate all Groups which have that favorite team words (like Dallas or Mavericks) in the Group name, OR locate nearby groups via zipcode proximity.
         
         
+        
+        
     }
     
 
@@ -52,7 +54,7 @@ class GroupSearchViewController: UIViewController, UITextFieldDelegate {
     override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
         
         if let groupExistVC = segue.destination as? GroupsExistViewController,
-            let zipcode = Int(zipcodeTextField.text!) {
+            let zipcode = Int16(zipcodeTextField.text!) {
             
             groupExistVC.favoriteTeam = favoriteTeamTextField.text
             groupExistVC.zipcode = zipcode
